+++ src/client_creds.rs
@@ -52,21 +52,14 @@ impl ClientCredentialsSpotify {
 
     /// Obtains the client access token for the app without saving it into the
     /// cache file. The resulting token is saved internally.
+    // TODO: handle with and without cache
     #[maybe_async]
-    pub async fn request_client_token_without_cache(&mut self) -> ClientResult<()> {
+    pub async fn request_token(&mut self) -> ClientResult<()> {
         let mut data = Form::new();
         data.insert(headers::GRANT_TYPE, headers::GRANT_CLIENT_CREDS);
 
         self.token = Some(self.fetch_access_token(&data).await?);
 
         Ok(())
-    }
-
-    /// The same as `request_client_token_without_cache`, but saves the token
-    /// into the cache file if possible.
-    #[maybe_async]
-    pub async fn request_client_token(&mut self) -> ClientResult<()> {
-        self.request_client_token_without_cache().await?;
-        self.write_token_cache()
     }
 }
